+++ onjung-chatting/src/main/java/com/onjung/onjung/repository/ChatRepository.java
@@ -1,0 +1,7 @@
+package com.onjung.onjung.repository;
+
+import com.onjung.onjung.domain.Chat;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ChatRepository extends JpaRepository<Chat, String> {
+}
