+++ onjung-chatting/src/main/java/com/onjung/onjung/repository/RoomRepository.java
@@ -3,7 +3,6 @@ package com.onjung.onjung.repository;
 import com.onjung.onjung.domain.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoomService extends JpaRepository<Room, Long> {
-
+public interface RoomRepository extends JpaRepository<Room, Long> {
 
 }
